processing/requirements.txt
connexion
PyYAML
connexion[swagger-ui]
pymongo
mysql-connector
pymongo